Check Content-Length against minSize and passthrough writes if no gzip (#71)
Currently there's no way to generate a HEAD response with the correct
headers as the GET unless you set the minSize as 0, but then gzip headers
will be written in Close. Instead, allow a Write(nil) that will set the
correct headers based on the Content-Length/Content-Type headers and only
initialize a writer if there is a non-zero-length Write. If the
Content-Length cannot be determined, you cannot generate the response
because it cannot know if minSize would've been met.

Additionally, if we determined that the request should not be compressed
we should passthrough writes immediately rather than waiting until
Close.

Fixes #70
Fixes #64