Start plain response as soon as it's certain
When we get the response headers it might be pretty
much already known whether we will have to serve plain
content instead of gzipping. That is either when
the content-length header is set with a small enough size,
there is already an encoding in the response or the
content-type is already set to a content we don't handle.